+++ src/components/stopPoster/footer.js
