+++ src/components/stopPoster/header.css
