+++ src/components/stopPoster/info.css
