+++ src/components/stopPoster/stopPoster.css
